% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_uncomment_chunk.R
\name{comment_chunk_addin}
\alias{comment_chunk_addin}
\title{Comment out the highlighted code chunk}
\usage{
comment_chunk_addin()
}
\description{
Call this as an addin to comment out selected rows by inserting a pound # 
in front of the row
}

