% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_uncomment_chunk.R
\name{identify_comment_rows}
\alias{identify_comment_rows}
\title{a function that identify which selected rows are/are not commment rows}
\usage{
identify_comment_rows(context_obj)
}
\description{
a function that identify which selected rows are/are not commment rows
}

