% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_uncomment_chunk.R
\name{uncomment_chunk_addin}
\alias{uncomment_chunk_addin}
\title{Uncomment the highlighted code chunk}
\usage{
uncomment_chunk_addin()
}
\description{
Call this as an addin to uncomment the selected rows by deleting the leading 
pound #. Selected rows that were originally not commented out will be ignored.
}

